netlify-published-dateプラグインのcontentsConverterオプションに指定する関数を外部ファイル化
