+++ index.js
@@ -1,5 +1,4 @@
 const netlifyPublishedDate = require('@sounisi5011/metalsmith-netlify-published-date');
-const cheerio = require('cheerio');
 const Metalsmith = require('metalsmith');
 const assetsConvention = require('metalsmith-assets-convention');
 const collections = require('metalsmith-collections');
@@ -11,6 +10,9 @@ const {
   render: pugRender,
 } = require('metalsmith-pug-extra');
 
+const {
+  ignoreContentsEquals,
+} = require('./src/plugin-options/netlify-published-date');
 const anotherSource = require('./src/plugins/another-source');
 const blankshield = require('./src/plugins/blankshield');
 const commentFrontmatter = require('./src/plugins/comment-matters');
@@ -101,79 +103,7 @@ Metalsmith(__dirname)
   )
   .use(
     netlifyPublishedDate({
-      contentsConverter(contents) {
-        /**
-         * @see https://github.com/sounisi5011/metalsmith-netlify-published-date/blob/v0.1.0/example/remove-time-elem.js
-         */
-
-        const netlifyDeployUrlRegExp = /^https?:\/\/[0-9a-f]+--[0-9a-z-]+\.netlify\.com(?=[/?#]|$)/i;
-        let isUpdated = false;
-
-        try {
-          const $ = cheerio.load(contents.toString());
-
-          // canonical URLを置換
-          $('head link[rel~=canonical][href^="http"]').each(
-            (index, element) => {
-              const $meta = $(element);
-              const hrefAttr = $meta.attr('href');
-              if (netlifyDeployUrlRegExp.test(hrefAttr)) {
-                $meta.attr(
-                  'href',
-                  hrefAttr.replace(netlifyDeployUrlRegExp, process.env.URL),
-                );
-                isUpdated = true;
-              }
-            },
-          );
-
-          // OGPの絶対URLを置換
-          $(
-            'head meta:matches([property^="og:"], [property^="twitter:"])[content^="http"]',
-          ).each((index, element) => {
-            const $meta = $(element);
-            const contentAttr = $meta.attr('content');
-            if (netlifyDeployUrlRegExp.test(contentAttr)) {
-              $meta.attr(
-                'content',
-                contentAttr.replace(netlifyDeployUrlRegExp, process.env.URL),
-              );
-              isUpdated = true;
-            }
-          });
-
-          // itemprop属性を持つtime要素を置換
-          const $timeListMap = new Map();
-          $('time[itemprop~=datePublished],time[itemprop~=dateModified]').each(
-            (index, element) => {
-              const $time = $(element);
-              const scopeDepth = $time.parents('[itemscope]').length;
-
-              if (!$timeListMap.has(scopeDepth)) {
-                $timeListMap.set(scopeDepth, new Set());
-              }
-
-              $timeListMap.get(scopeDepth).add($time);
-            },
-          );
-          const minScopeDepth = Math.min(...$timeListMap.keys());
-          $timeListMap.get(minScopeDepth).forEach($time => {
-            $time.empty();
-            if ($time.is('[datetime]')) {
-              $time.attr('datetime', '');
-            }
-            isUpdated = true;
-          });
-
-          if (isUpdated) {
-            return Buffer.from($.html());
-          }
-        } catch (err) {
-          //
-        }
-
-        return contents;
-      },
+      contentsConverter: ignoreContentsEquals,
       plugins: [
         pugRender({
           locals: {
